refactor: move chat feature files
