+++ src/App.vue
@@ -28,8 +28,8 @@ import { EncryptionService } from "./services/encryption/EncryptionService"
 import { createKeplerWallet } from "./services/kepler/KeplerWallet"
 import { useAssistantsStore } from "./stores/assistants"
 import { useAuthStore } from "./stores/auth"
-import { useChatMessagesStore } from "./stores/chat-messages"
-import { useChatsStore } from "./stores/chats"
+import { useChatMessagesStore } from "./features/chats/stores/chat-messages"
+import { useChatsStore } from "./features/chats/stores/chats"
 import { useDialogsStore } from "./stores/dialogs"
 import { usePluginsStore } from "./stores/plugins"
 import { getMnemonic } from "./stores/tauri-store"
