+++ src/features/chats/components/ChatList.vue
@@ -55,13 +55,13 @@
 <script setup lang="ts">
 import { useQuasar } from "quasar"
 import { useListenKey } from "src/composables/listen-key"
-import { useChatsStore } from "src/stores/chats"
+import { useChatsStore } from "src/features/chats/stores/chats"
 import { useUserStore } from "src/stores/user"
 import { useUserPerfsStore } from "src/stores/user-perfs"
 import { isPlatformEnabled } from "src/utils/functions"
 import { ref, toRef } from "vue"
 import { useRouter } from "vue-router"
-import { useWorkspaceChats } from "../../composables/chats/useWorkspaceChats"
+import { useWorkspaceChats } from "../composables/useWorkspaceChats"
 import ChatListItem from "./ChatListItem.vue"
 import SearchChats from "./SearchChats.vue"
 import UserListDialog from "./UserListDialog.vue"
