+++ src/features/chats/components/ChatListItem.vue
@@ -60,9 +60,9 @@ import AAvatar from "src/components/AAvatar.vue"
 import { dialogOptions } from "src/utils/values"
 import { computed, ref, toRef } from "vue"
 import { useRoute } from "vue-router"
-import { useWorkspaceChats } from "../../composables/chats/useWorkspaceChats"
-import MenuButton from "../ExpansionItem/MenuButton.vue"
-import MenuItem from "../MenuItem.vue"
+import { useWorkspaceChats } from "../composables/useWorkspaceChats"
+import MenuButton from "src/components/ExpansionItem/MenuButton.vue"
+import MenuItem from "src/components/MenuItem.vue"
 import { ChatMapped } from "@/services/supabase/types"
 
 const props = defineProps<{
