+++ src/features/chats/components/SearchChats.vue
@@ -84,7 +84,7 @@
 <script setup lang="ts">
 import Mark from "mark.js"
 import { QList, useDialogPluginComponent } from "quasar"
-import { useChatsStore } from "src/stores/chats"
+import { useChatsStore } from "src/features/chats/stores/chats"
 import { escapeRegex } from "src/utils/functions"
 import { nextTick, watch, ref, watchEffect } from "vue"
