+++ src/features/chats/stores/chats.ts
@@ -1,9 +1,9 @@
 import { throttle } from "lodash"
 import { defineStore } from "pinia"
-import { useChatsWithSubscription } from "src/composables/chats/useChatsWithSubscription"
+import { useChatsWithSubscription } from "src/features/chats/composables/useChatsWithSubscription"
 import { supabase } from "src/services/supabase/client"
 import { readonly } from "vue"
-import { useUserStore } from "./user"
+import { useUserStore } from "src/stores/user"
 import { ChatMapped } from "@/services/supabase/types"
 
 export const useChatsStore = defineStore("chats", () => {
