+++ src/features/chats/views/ChatSettings.vue
@@ -96,9 +96,9 @@ import NotificationPanel from "src/components/common/NotificationPanel.vue"
 import PickAvatarDialog from "src/components/PickAvatarDialog.vue"
 import ViewCommonHeader from "src/components/ViewCommonHeader.vue"
 
-import { useIsChatAdmin } from "src/composables/chats/useChatAdmin"
+import { useIsChatAdmin } from "src/features/chats/composables/useChatAdmin"
 import { syncRef } from "src/composables/sync-ref"
-import { useChatsStore } from "src/stores/chats"
+import { useChatsStore } from "src/features/chats/stores/chats"
 import { useWorkspacesStore } from "src/stores/workspaces"
 import { pageFhStyle } from "src/utils/functions"
 import { computed, ref, toRaw, watch } from "vue"
