+++ src/layouts/PinnedChats.vue
@@ -31,7 +31,7 @@
 <script setup lang="ts">
 import { storeToRefs } from "pinia"
 import AAvatar from "src/components/AAvatar.vue"
-import { useChatsStore } from "src/stores/chats"
+import { useChatsStore } from "src/features/chats/stores/chats"
 import { useWorkspacesStore } from "src/stores/workspaces"
 import { computed } from "vue"
 import { useRouter } from "vue-router"
