+++ src/router/routes.ts
@@ -3,15 +3,15 @@ import ErrorNotFound from "pages/ErrorNotFound.vue"
 import { i18n } from "src/boot/i18n"
 import AccountPage from "src/pages/AccountPage.vue"
 import AssistantsPage from "src/pages/AssistantsPage.vue"
-import ChatsPage from "src/pages/ChatsPage.vue"
+import ChatsPage from "src/features/chats/pages/ChatsPage.vue"
 import EmptyPage from "src/pages/EmptyPage.vue"
 import PluginsPage from "src/pages/PluginsPage.vue"
 import SetProvider from "src/pages/SetProvider.vue"
 import WorkspacePage from "src/pages/WorkspacePage.vue"
 import AssistantsMarket from "src/views/AssistantsMarket.vue"
 import AssistantView from "src/views/AssistantView.vue"
-import ChatSettings from "src/views/ChatSettings.vue"
-import ChatView from "src/views/ChatView.vue"
+import ChatSettings from "src/features/chats/views/ChatSettings.vue"
+import ChatView from "src/features/chats/views/ChatView.vue"
 import CustomProvider from "src/views/CustomProvider.vue"
 import DialogView from "src/views/DialogView.vue"
 import PluginAdjust from "src/views/PluginAdjust.vue"
